requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
PyNaCl==1.3.0
pandas
dnspython==1.16.0
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord==1.0.1
discord.py==1.5.0
idna==2.10
multidict==4.7.6
pip==20.2.3
python-dotenv==0.14.0
setuptools==50.3.0	
yarl==1.5.1